pyproject.toml
[project]
name = "mix-sample"
version = "0.2.0"
description = "動画処理・合成・連結ライブラリ"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pillow>=11.3.0",
]

[project.scripts]
video-mixer = "video_mixer:main"
video-concatenator = "video_concatenator:main"
advanced-video-concatenator = "advanced_video_concatenator:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "video_mixer.py",
    "video_concatenator.py", 
    "advanced_video_concatenator.py",
    "video_processing_lib.py",
    "__init__.py"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.2",
    "pytest-cov>=6.3.0",
    "ruff>=0.12.12",
]
